accelerators: (rescale_up_extension) Add rescale up extension (#551)
* add rescale up

* change pixi and format

* remove unneccesary file

* fix issues

* format files

* merge rescale extensions

* only keep 1 rescale folder

* format files

* remove from ci

* delete unnecessary files

* remove print when correct

* change verify functions

+++ tests/filecheck/dialects/kernel/ops.mlir
@@ -24,6 +24,6 @@
 // CHECK-NEXT:   %4 = kernel.qmac %0, %0 zp_lhs : %0 zp_rhs : %0 : i32, i32, i32, i32 -> i32
 // CHECK-GENERIC-NEXT:   %4 = "kernel.qmac"(%0, %0, %0, %0) : (i32, i32, i32, i32) -> i32
 
-%5 = "kernel.rescale"(%0) {input_zp = 23 : i8, output_zp = -23 : i8, multiplier = array<i32: 12345>, shift = array<i8: 30>, min_int = -128 : i8, max_int = 127 : i8, double_round = true} : (i32) -> i32
-// CHECK-NEXT:   %5 = kernel.rescale %0 {input_zp = 23 : i8, output_zp = -23 : i8, multiplier = array<i32: 12345>, shift = array<i8: 30>, min_int = -128 : i8, max_int = 127 : i8, double_round = true} : (i32) -> i32
-// CHECK-GENERIC-NEXT    %5 = "kernel.rescale"(%0) {input_zp = 23 : i8, output_zp = -23 : i8, multiplier = array<i32: 12345>, shift = array<i8: 30>, min_int = -128 : i8, max_int = 127 : i8, double_round = true} : (i32) -> i32
+%5 = "kernel.rescale"(%0) {input_zp = 23 : i32, output_zp = -23 : i32, multiplier = array<i32: 12345>, shift = array<i8: 30>, min_int = -128 : i32, max_int = 127 : i32, double_round = true} : (i32) -> i32
+// CHECK-NEXT:   %5 = kernel.rescale %0 {input_zp = 23 : i32, output_zp = -23 : i32, multiplier = array<i32: 12345>, shift = array<i8: 30>, min_int = -128 : i32, max_int = 127 : i32, double_round = true} : (i32) -> i32
+// CHECK-GENERIC-NEXT    %5 = "kernel.rescale"(%0) {input_zp = 23 : i32, output_zp = -23 : i32, multiplier = array<i32: 12345>, shift = array<i8: 30>, min_int = -128 : i32, max_int = 127 : i32, double_round = true} : (i32) -> i32
